+++ compiler/backend/src/test/java/org/apache/nemo/compiler/backend/nemo/StagePartitionerTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.nemo.runtime.common.plan;
+package org.apache.nemo.compiler.backend.nemo;
 
 import org.apache.nemo.common.dag.DAGBuilder;
 import org.apache.nemo.common.ir.IRDAG;
